tests/submission/AryehGregor/base64.html

<!doctype html>
<title>atob()/btoa() tests</title>
<meta charset=utf-8>
<div id=log></div>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>
function mybtoa(s) {
	var out = "";
	for (var i = 0; i < s.length; i += 3) {
		out += btoaLookup((s.charCodeAt(i) & 0xfc) >> 2);
		out += btoaLookup( ((s.charCodeAt(i) & 0x03) << 4) | ((s.charCodeAt(i + 1) & 0xf0) >> 4) );
		if (i + 1 < s.length) {
			out += btoaLookup( ((s.charCodeAt(i + 1) & 0x0f) << 2) | ((s.charCodeAt(i + 2) & 0xc0) >> 6) );
		} else {
			out += '=';
		}
		if (i + 2 < s.length) {
			out += btoaLookup(s.charCodeAt(i + 2) & 0x3f);
		} else {
			out += '=';
		}
	}
	return out;
}

function btoaLookup(idx) {
	if (idx < 26) {
		return String.fromCharCode(idx + 'A'.charCodeAt(0));
	}
	if (idx < 52) {
		return String.fromCharCode(idx - 26 + 'a'.charCodeAt(0));
	}
	if (idx < 62) {
		return String.fromCharCode(idx - 52 + '0'.charCodeAt(0));
	}
	if (idx == 62) {
		return '+';
	}
	if (idx == 63) {
		return '/';
	}
	// Throw INVALID_CHARACTER_ERR exception here -- won't be hit in the tests.
}

function testBtoa(input) {
	for (var i = 0; i < input.length; i++) {
		if (input.charCodeAt(i) > 255) {
			assert_throws("INVALID_CHARACTER_ERR",
				function() { btoa(input) }
			);
		}
		return;
	}
	assert_equals(btoa(input), mybtoa(input));
}

var tests = [];
for (var i = 0; i < 258; i++) {
	tests.push(["Correct btoa output for code point " + i, String.fromCharCode(i)]);
}
tests.push(["Correct btoa output for code point 10000", String.fromCharCode(10000)]);
tests.push(["Correct btoa output for code point 65534", String.fromCharCode(65534)]);
tests.push(["Correct btoa output for code point 65535", String.fromCharCode(65535)]);

var everything = "";
for (var i = 0; i < 256; i++) {
	everything += String.fromCharCode(i);
}
tests.push(["Correct btoa output for first 256 characters concatenated", everything]);

tests.push(["Correct btoa output for עברית", "עברית"]);
tests.push(["Correct btoa output for empty string", ""]);
tests.push(["Correct btoa output for a", "a"]);
tests.push(["Correct btoa output for ab", "ab"]);
tests.push(["Correct btoa output for abc", "abc"]);
tests.push(["Correct btoa output for abcd", "abcd"]);
tests.push(["Correct btoa output for abcde", "abcde"]);

tests.push(["Correct btoa output for code point 0x10001", String.fromCharCode(0xd840, 0xdc00)]);

generate_tests(testBtoa, tests);
</script>
